docker-aur-xen: build Archlinux Xen packages
Builds Archlinux Xen packages.  The packages will be available under pkg/ after
building & running this container.

Requires first using my docker-archlinux-base project to create the
archlinux-base-multilib image in the user's local Docker repository.  Uses a
tmpfs volume for the source dir.  Unfortunately, it's not currently possible to
use a tmpfs volume for the work dir because Docker mounts tmpfs volumes noexec
inside the container with no way to override it
(see: https://github.com/moby/moby/issues/32131).  The makepkg(1) patch is
necessary to avoid unnecessarily chmod'ing the src dir, which fails otherwise
because the tmpfs volume is also owned by root:root with no way to override it,
either.

NOTE: the makepkg(1) step currently *does not* verify the signature of the
downloaded Xen sources because I couldn't figure out an easy way to
automatically parse PKGBUILD to get the URL of the signature file & import
the signing key.  Since the build explicitly depends on python2.7 the simplest
thing will probably be to ADD and RUN some custom Python script.

Signed-off-by: Kent R. Spillner <kspillner@acm.org>